Speed up conversion of definitions
Running the pre-commit hooks one by one of the files seemed to cause an overhead of about 6 seconds per file or 132 seconds for converting and updating 29 files. By running the pre-commit hooks on them only once we can shorten down this to only one invocation and thus process the same 29 files in around 7 seconds in total.