Add versions to definitions (#21)
The DataProductDefinition now has a version field for the version number of the definition.

The validator ensures the filename has a version matching the version in the definition and that the test & draft definitions have no version in the filename and the version within the definition is 0.0.x. Definitions outside test & draft are on the other hand required to have a version in the filename (of the form _v{MAJOR}.{MINOR} with optional pre-release data) and it must match the full version ({MAJOR}.{MINOR}.{PATCH} with optional pre-release data) within the definition.