Update dependencies (#33)
Mainly updates FastAPI (and switches from `fastapi-slim` to `fastapi`, the latter was slimmed down as part of the 0.112.0 release, so no point in using the `-slim` version any more). Main reason for updating FastAPI is to be able to update starlette to a version where a CVE is fixed.

Also updated other dependencies to latest versions at the same time.