Remove headers related to authorization and consent (#39)
This PR does by default change the tooling so that the `Authorization`, `X-Authorization-Provider` and `X-Consent-Token` are by default excluded from the generated definitions and not required to be present in the definitions.

For data sharing services where the headers are still desired, the flags `--authorization-headers` and `--consent-headers` can be used in the converter to ensure it generates them and in the validator to ensure they exist in the definitions.